Conflicts:
	MyGraphics - SP2/MyGraphics/Source/SP2.cpp

+++ MyGraphics - SP2/MyGraphics/Source/MazeInteraction.h
@@ -29,8 +29,8 @@ public:
 	float gridSizeY;
 	int rowNumber;
 	int colNumber;
-	int startingPointX;
-	int startingPointY;
+	float startingPointX;
+	float startingPointY;
 
 	int mapLayout[10][10];
 private:
